Move suggestions service to own module
