+++ settings.gradle
@@ -207,7 +207,10 @@ include 'stroom-proxy:stroom-proxy-remote-api'
 include 'stroom-query:stroom-expression'
 include 'stroom-query:stroom-query-api'
 include 'stroom-query:stroom-query-common'
-include 'stroom-query:stroom-query-impl'
+
+// Suggestions
+include 'stroom-suggestions:stroom-suggestions-api'
+include 'stroom-suggestions:stroom-suggestions-impl'
 
 // Resource store
 include 'stroom-resource:stroom-resource-api'
@@ -273,3 +276,15 @@ include 'stroom-util-shared'
 
 
 include 'stroom-ssl-test'
+include 'stroom-suggestions'
+include 'stroom-suggestions:stroom-suggestions-api'
+findProject(':stroom-suggestions:stroom-suggestions-api')?.name = 'stroom-suggestions-api'
+include 'stroom-suggestions:stroom-suggestions-impl'
+findProject(':stroom-suggestions:stroom-suggestions-impl')?.name = 'stroom-suggestions-impl'
+include 'stroom-suggestions:stroom-suggestions-impl:main'
+findProject(':stroom-suggestions:stroom-suggestions-impl:main')?.name = 'main'
+include 'stroom-suggestions:stroom-suggestions-impl'
+findProject(':stroom-suggestions:stroom-suggestions-impl')?.name = 'stroom-suggestions-impl'
+include 'stroom-suggestions:stroom-suggestions-impl:main'
+findProject(':stroom-suggestions:stroom-suggestions-impl:main')?.name = 'main'
+
