+++ stroom-app/build.gradle
@@ -163,7 +163,8 @@ dependencies {
     implementation project(':stroom-query:stroom-expression')
     implementation project(':stroom-query:stroom-query-api')
     implementation project(':stroom-query:stroom-query-common')
-    implementation project(':stroom-query:stroom-query-impl')
+    implementation project(':stroom-suggestions:stroom-suggestions-api')
+    implementation project(':stroom-suggestions:stroom-suggestions-impl')
     implementation project(':stroom-receive:stroom-receive-common')
     implementation project(':stroom-receive:stroom-receive-rules-impl')
     implementation project(':stroom-resource:stroom-resource-api')
