+++ stroom-meta/stroom-meta-impl/src/main/java/stroom/meta/impl/MetaModule.java
@@ -7,8 +7,8 @@ import stroom.meta.api.MetaService;
 import stroom.meta.api.PhysicalDelete;
 import stroom.meta.shared.Meta;
 import stroom.meta.shared.MetaFields;
-import stroom.query.impl.SuggestionsServiceBinder;
 import stroom.searchable.api.Searchable;
+import stroom.suggestions.api.SuggestionsServiceBinder;
 import stroom.util.RunnableWrapper;
 import stroom.util.guice.GuiceUtil;
 import stroom.util.guice.RestResourcesBinder;
