+++ stroom-search/stroom-search-elastic/src/main/java/stroom/search/elastic/ElasticSearchModule.java
@@ -21,13 +21,13 @@ import stroom.docstore.api.DocumentActionHandlerBinder;
 import stroom.explorer.api.ExplorerActionHandler;
 import stroom.importexport.api.ImportExportActionHandler;
 import stroom.job.api.ScheduledJobsBinder;
-import stroom.query.impl.SuggestionsServiceBinder;
 import stroom.search.elastic.indexing.ElasticIndexingElementModule;
 import stroom.search.elastic.search.ElasticIndexQueryResourceImpl;
 import stroom.search.elastic.shared.ElasticClusterDoc;
 import stroom.search.elastic.shared.ElasticIndexDoc;
 import stroom.search.elastic.suggest.ElasticSuggestionsQueryHandler;
 import stroom.search.elastic.suggest.ElasticSuggestionsQueryHandlerImpl;
+import stroom.suggestions.api.SuggestionsServiceBinder;
 import stroom.util.RunnableWrapper;
 import stroom.util.entityevent.EntityEvent;
 import stroom.util.guice.GuiceUtil;
