+++ stroom-search/stroom-search-elastic/src/main/java/stroom/search/elastic/suggest/ElasticSuggestionsQueryHandler.java
@@ -1,6 +1,6 @@
 package stroom.search.elastic.suggest;
 
-import stroom.query.impl.SuggestionsQueryHandler;
+import stroom.suggestions.api.SuggestionsQueryHandler;
 
 public interface ElasticSuggestionsQueryHandler extends SuggestionsQueryHandler {
