+++ stroom-suggestions/stroom-suggestions-api/src/main/java/stroom/suggestions/api/SuggestionsService.java
@@ -1,4 +1,4 @@
-package stroom.query.impl;
+package stroom.suggestions.api;
 
 import stroom.query.shared.FetchSuggestionsRequest;
