+++ stroom-suggestions/stroom-suggestions-api/src/main/java/stroom/suggestions/api/SuggestionsServiceBinder.java
@@ -1,4 +1,4 @@
-package stroom.query.impl;
+package stroom.suggestions.api;
 
 import com.google.inject.Binder;
 import com.google.inject.multibindings.MapBinder;
